Add support for connecting to the Band via Bluetooth
This is a big one that was hard to split, so it's coming in at once.
Here's a list of the changes:

* Added the ability to discover and connect to devices over BT. They
need to be paired to the PC first and a USB connection is always
preferred (i.e. unBand will first search for USB connections then BT)
* Getting BT was ugly, it required a lot of digging in the debugger to
figure out various interfaces / objects that are not normally exposed.
To get to those internals we have gone down the horrible path of
modifying the Cargo client dll on the fly (see the new CargoDll class)
to flip some fields from private to public. This is extremely dangerous
and unstable, but then again, so is support for BT in general. This
should not affect regular USB usage.
* Added Mono.Cecil binaries which are used for the on-the-fly rewrite
* Added a bunch of "extension" classes in a new project,
unBand.CargoClientExtender, which facilitate Bluetooth discovery and
connection based on the WinRT BT APIs. This will require Windows 8.1+ -
we'll need to put in a "do not use" solution for Windows 7 later (if
needed).
** Note: These classes are expected to be buggy, most of the code is
based on guess work. I've documented various assumptions in the
comments.
* Initialization of the various Band helpers (Theme, Tiles etc) has
moved out of the constructor and into an InitAsync method, which helps
reduce load on a Band connected via Bluetooth.
* fixed #25 since I was using a newly reset Band that was causing
crashes