Added bootstrapper for building against a modified CargoClient dll
Since we link against the modified CargoClient at build time we need to
have the dll included with the project, but we obviously don't want to
distribute it with the code. The solution is to run the bootstrapper
which simply creates the modified dll and dumps it into convenient
locations for VS to pick up during build (this location is ignored in
.gitignore). As part of this change we split out all editing of
CargoClient to a separate dll, which is then referenced by a new
project, unBand.dev.Bootstrap.

This is not an ideal solution but it should suffice for our needs.